+++ stts/macos/stts/Sources/stts/stt/Stt.swift
@@ -13,16 +13,19 @@ class SttRecognitionOptions {
   let punctuation: Bool
   let contextualStrings: [String]
   let taskHint: SFSpeechRecognitionTaskHint?
+  let offline: Bool
 
-  init(punctuation: Bool, contextualStrings: [String], taskHint: SFSpeechRecognitionTaskHint? = nil) {
+  init(punctuation: Bool, contextualStrings: [String], taskHint: SFSpeechRecognitionTaskHint?, offline: Bool) {
     self.punctuation = punctuation
     self.contextualStrings = contextualStrings
     self.taskHint = taskHint
+    self.offline = offline
   }
   
   static func fromMap(_ map: [String: Any]) -> SttRecognitionOptions {
     let punctuation = map["punctuation"] as? Bool ?? false
     let contextualStrings = map["contextualStrings"] as? [String] ?? []
+    let offline = map["offline"] as? Bool ?? true
     
     var taskHint: SFSpeechRecognitionTaskHint?
     if let macosOptions = map["macos"] as? [String : Any] {
@@ -39,7 +42,8 @@ class SttRecognitionOptions {
     return SttRecognitionOptions(
       punctuation: punctuation,
       contextualStrings: contextualStrings,
-      taskHint: taskHint
+      taskHint: taskHint,
+      offline: offline
     )
   }
 }
@@ -51,6 +55,7 @@ class Stt {
   private var recognizer: SFSpeechRecognizer?
   private var recognitionRequest: SFSpeechAudioBufferRecognitionRequest?
   private var recognitionTask: SFSpeechRecognitionTask?
+  private var stopTimer: Timer?
   
   private var stateEventHandler: SttStateStreamHandler
   private var resultEventHandler: SttResultStreamHandler
@@ -107,6 +112,8 @@ class Stt {
   }
   
   func stop() {
+    stopTimer?.invalidate()
+
     recognitionRequest?.endAudio()
     recognitionRequest = nil
     
@@ -139,7 +146,7 @@ class Stt {
     
     recognitionRequest.shouldReportPartialResults = true
     
-    if recognizer.supportsOnDeviceRecognition {
+    if options.offline && recognizer.supportsOnDeviceRecognition {
       recognitionRequest.requiresOnDeviceRecognition = true
     }
 
@@ -170,6 +177,11 @@ class Stt {
         }
         if isFinal {
           self.stop()
+        } else if !recognitionRequest.requiresOnDeviceRecognition {
+          self.stopTimer?.invalidate()
+          self.stopTimer = Timer.scheduledTimer(withTimeInterval: 3, repeats: false) { [weak self] timer in
+            self?.stop()
+          }
         }
       } else if self.recognitionTask != nil, let error = error { // check task to not fire error event on stop/cancel
         self.stateEventHandler.sendErrorEvent(error)