+++ stts/CHANGELOG.md
@@ -1,4 +1,5 @@
-## next
+## 1.2.0
+* feat(STT): Allow online usage.
 * fix(iOS/macOS): compilation issue under some circumstances.
 * fix(iOS/macOS/TTS): Event firing when stressing the plugin.
 * fix(iOS/macOS/TTS): Stay away from UI thread.